requirements.txt
# Python 3.11

# Core numeric / tensor
numpy==1.26.4
torch==2.10.0
torchvision==0.25.0

# Vision models
ultralytics==8.4.24
opencv-python==4.9.0.80
Pillow==12.1.1
clip @ git+https://github.com/openai/CLIP.git@ded190a052fdf4585bd685cee5bc96e0310d2c93
insightface==0.7.3
onnxruntime==1.24.4
easyocr==1.7.2

# Utilities
PyYAML==6.0.2
tqdm==4.67.1
gdown==5.2.1
